[Serving] Switching to global request callback
Prior to this PR, each request has its own callback function.
This PR switches to using a global callback function that all
requests share, so that the conversion of Python function
to PackedFunc only happens once.

Additionally, this PR introduces the use of `call_soon_threadsafe`
method of asyncio in AsyncEngine.