[Fix] Fix the two-stage softmax func by removing log2e
When two-stage softmax was introduced, we use a log2e numeric
transformation for some potentially better performance.

However, under the case of low temperature, the log2e transformation
is not numerically stable, which may cause the softmax result not
summing up to 1.

This PR fixes this by removing all the log2e related calculation.